+++ Luo Painter/$DrawPage.Option/DrawPage.Effect.DiscreteTransfer.cs
@@ -16,7 +16,7 @@
             };
         }
 
-        private void SetDiscreteTransfer(NumberPickerMode mode, int e)
+        private void SetDiscreteTransfer(NumberPickerMode mode, float e)
         {
         }
